인터페이스 구현체 형식으로 변경
