+++ src/main/java/goorm/back/zo6/user/application/UserService.java
@@ -1,55 +1,9 @@
 package goorm.back.zo6.user.application;
 
-import goorm.back.zo6.auth.util.JwtUtil;
-import goorm.back.zo6.common.exception.CustomException;
-import goorm.back.zo6.common.exception.ErrorCode;
-import goorm.back.zo6.user.dto.request.LoginRequest;
-import goorm.back.zo6.user.dto.request.SignUpRequest;
-import goorm.back.zo6.user.dto.response.LoginResponse;
-import goorm.back.zo6.user.dto.response.SignUpResponse;
-import goorm.back.zo6.user.dto.response.UserResponse;
-import goorm.back.zo6.user.domain.Role;
-import goorm.back.zo6.user.domain.User;
-import goorm.back.zo6.user.domain.UserRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@RequiredArgsConstructor
-@Log4j2
-@Service
-@Transactional(readOnly = true)
-public class UserService {
-    private final UserRepository userRepository;
-    private final PasswordEncoder passwordEncoder;
-
-    public UserResponse findById(Long userId){
-        User user = userRepository.findById(userId).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
-        return UserResponse.from(user);
-    }
-
-    @Transactional
-    public SignUpResponse signUp(SignUpRequest request){
-        userRepository.findByEmail(request.email())
-                .ifPresent(user -> {
-                    throw new CustomException(ErrorCode.USER_ALREADY_EXISTS);
-                });
-
-        User user = userRepository.save(User.singUpUser(request.email(),request.name(), passwordEncoder.encode(request.password()), request.phone(), request.birth_date(), Role.of("USER")));
-
-        return SignUpResponse.from(user);
-    }
-
-    public UserResponse findByToken(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
-        return UserResponse.from(user);
-    }
-
-    @Transactional
-    public void delete(String email) {
-        User user = userRepository.findByEmail(email).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
-        userRepository.deleteById(user.getId());
-    }
+public interface UserService {
+    UserResponse findById(Long userId);
+    SignUpResponse signUp(SignUpRequest request);
+    UserResponse findByToken(String email);
+    void delete(String email);
+    UserResponse addPhoneNumber(String email, String phoneNumber);
 }
