+++ src/main/java/goorm/back/zo6/user/application/UserServiceImpl.java
@@ -1,0 +1,70 @@
+package goorm.back.zo6.user.application;
+
+import goorm.back.zo6.common.exception.CustomException;
+import goorm.back.zo6.common.exception.ErrorCode;
+import goorm.back.zo6.user.domain.Role;
+import goorm.back.zo6.user.domain.User;
+import goorm.back.zo6.user.domain.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Log4j2
+@Service
+@Transactional(readOnly = true)
+public class UserServiceImpl implements UserService {
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @Override
+    public UserResponse findById(Long userId){
+        User user = userRepository.findById(userId).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+        return UserResponse.from(user);
+    }
+
+    @Override
+    @Transactional
+    public SignUpResponse signUp(SignUpRequest request){
+        userRepository.findByEmail(request.email())
+                .ifPresent(user -> {
+                    throw new CustomException(ErrorCode.USER_ALREADY_EXISTS);
+                });
+
+        User user = userRepository.save(User.singUpUser(
+                request.email(),
+                request.name(),
+                passwordEncoder.encode(request.password()),
+                request.phone(),
+                request.birth_date(),
+                Role.of("USER")
+        ));
+
+        return SignUpResponse.from(user);
+    }
+
+    @Override
+    public UserResponse findByToken(String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+        return UserResponse.from(user);
+    }
+
+    @Override
+    @Transactional
+    public void delete(String email) {
+        User user = userRepository.findByEmail(email).orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+        userRepository.deleteById(user.getId());
+    }
+
+    @Override
+    @Transactional
+    public UserResponse addPhoneNumber(String email, String phoneNumber) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
+
+        user.updatePhoneNumber(phoneNumber);
+        return UserResponse.from(user);
+    }
+}
